Implement corrected cohesion review repairs
