+++ installer/planning.py
@@ -12,7 +12,6 @@ from .foundation import (
     CODEX_CLI_CANDIDATES,
     GROK_CLI_CANDIDATES,
     PROFILE_ROLES,
-    READER_ROOT,
     REPO_ROOT,
     SHARED_ADAPTER_NAMES,
     _bin_dir,
@@ -66,6 +65,11 @@ from .packages import (
     template_adapter_body,
 )
 from .runtime import private_runtime_action
+from .planning_support import (
+    _frontend_plan,
+    _reader_payload_files,
+    _script_source,
+)
 from .hosts import host_item_path, load_host_adapters, preflight_instruction_target
 
 def detect_hosts(home: Path | None = None) -> tuple[bool, bool, bool]:
@@ -150,28 +154,18 @@ def _build_user_plan(
     notices = REPO_ROOT / "THIRD_PARTY_NOTICES.md"
     if notices.is_file():
         lib_copies.append((notices, lib_home / notices.name))
-    reader_runtime_files = (READER_ROOT / "__init__.py", *sorted((READER_ROOT / "scripts").glob("*.py")), *(REPO_ROOT / "scripts" / name for name in "state_root.py tickets_adapters.py tickets_bound.py tickets_ceiling.py tickets_format.py tickets_lifecycle.py tickets_markdown.py tickets_readiness.py tickets_registry.py tickets_sequence.py tickets_shapes.py".split())); reader_manifest_files = sorted((READER_ROOT / "docs").glob("*.json"))
-    for path in (*reader_runtime_files, *reader_manifest_files):
+    for path in _reader_payload_files():
         if path.is_file():
             lib_copies.append((path, lib_home / path.relative_to(REPO_ROOT)))
-    frontend_source = READER_ROOT / "web" / "dist"
-    frontend_home = _frontend_home()
-    frontend_identity = _frontend_manifest_identity(frontend_source)
-    if frontend_identity is None:
-        raise RuntimeError("reader/web/dist is missing its immutable index.html distribution")
-    frontend_assets = [
-        (path, frontend_home / path.relative_to(frontend_source))
-        for path in sorted(frontend_source.rglob("*"))
-        if path.is_file()
-    ]
-    installed_frontend_identity = _frontend_manifest_identity(frontend_home)
-    frontend_action = (
-        "reuse"
-        if installed_frontend_identity == frontend_identity
-        else ("repair" if frontend_home.exists() else "create")
+    frontend_home, frontend_identity, frontend_assets, frontend_action = _frontend_plan(
+        _frontend_home, _frontend_manifest_identity
     )
     names = script_name_discoverer(REPO_ROOT / "scripts")
-    scripts = [(READER_ROOT / "scripts" / name if name == "ui.py" else REPO_ROOT / "scripts" / name, bin_dir / name) for name in names if (READER_ROOT / "scripts" / name if name == "ui.py" else REPO_ROOT / "scripts" / name).is_file()]
+    scripts = [
+        (_script_source(name), bin_dir / name)
+        for name in names
+        if _script_source(name).is_file()
+    ]
 
     claude_scope_home = _claude_scope_home("user", None)
     codex_user_home = _codex_user_home()
