+++ scripts/tickets_commands.py
@@ -15,6 +15,7 @@ if __package__:
     from .tickets_generations import GENERATION_SUBCOMMANDS
     from .tickets_attempts import DISPATCH_COMMIT_USAGE, DISPATCH_OPEN_USAGE, DISPATCH_REPLACE_USAGE, DISPATCH_RETIRE_USAGE
     from .tickets_dispatch_packet import DISPATCH_PACKET_USAGE, DISPATCH_RECEIVE_USAGE
+    from .tickets_dispatch_receipt import DISPATCH_RECEIPT_USAGE
     from .tickets_join import DISPATCH_JOIN_USAGE, DISPATCH_OUTCOME_USAGE
 else:
     from tickets_format import EXECUTOR_SECTIONS, TERMINAL_STATES, VALID_STATUSES, _read_utf8
@@ -33,6 +34,7 @@ else:
     _dispatch_packet = __import__("tickets_dispatch_packet")
     DISPATCH_PACKET_USAGE = _dispatch_packet.DISPATCH_PACKET_USAGE
     DISPATCH_RECEIVE_USAGE = _dispatch_packet.DISPATCH_RECEIVE_USAGE
+    DISPATCH_RECEIPT_USAGE = __import__("tickets_dispatch_receipt").DISPATCH_RECEIPT_USAGE
     _join = __import__("tickets_join")
     DISPATCH_JOIN_USAGE = _join.DISPATCH_JOIN_USAGE
     DISPATCH_OUTCOME_USAGE = _join.DISPATCH_OUTCOME_USAGE
@@ -69,6 +71,7 @@ SUBCOMMAND_USAGE = {
     "dispatch-join": DISPATCH_JOIN_USAGE,
     "dispatch-packet": DISPATCH_PACKET_USAGE,
     "dispatch-receive": DISPATCH_RECEIVE_USAGE,
+    "dispatch-receipt": DISPATCH_RECEIPT_USAGE,
     "check": CHECK_USAGE,
     "set-status": "set-status <run> <id> <status>",
     "join-noop-repair": JOIN_NOOP_REPAIR_USAGE,
@@ -99,6 +102,7 @@ SUBCOMMAND_SUMMARY = {
     "dispatch-join": "Commit or replay one outcome-fenced join and its lifecycle transition.",
     "dispatch-packet": "Commit or replay one reference or inline dispatch-v1 packet projection.",
     "dispatch-receive": "Validate one dispatch-v1 packet against its receipt identity and authority.",
+    "dispatch-receipt": "Read one durable accepted dispatch receipt without mutation.",
     "check": "Anchor one completed durable checker stage to its target's checked_by field.",
     "set-status": f"Set lifecycle status to one of {sorted(VALID_STATUSES)}.",
     "join-noop-repair": "Atomically attribute and complete a clean repair at the join without dispatch.",
