+++ scripts/tickets_dispatch.py
@@ -28,10 +28,12 @@ else:
 if __package__:
     from .tickets_attempts import _cmd_dispatch_commit, _cmd_dispatch_open, _cmd_dispatch_replace, _cmd_dispatch_retire
     from .tickets_dispatch_packet import _cmd_dispatch_packet, _cmd_dispatch_receive
+    from .tickets_dispatch_receipt import _cmd_dispatch_receipt
     from .tickets_join import _cmd_dispatch_join, _cmd_dispatch_outcome
 else:
     from tickets_attempts import _cmd_dispatch_commit, _cmd_dispatch_open, _cmd_dispatch_replace, _cmd_dispatch_retire
     from tickets_dispatch_packet import _cmd_dispatch_packet, _cmd_dispatch_receive
+    from tickets_dispatch_receipt import _cmd_dispatch_receipt
     from tickets_join import _cmd_dispatch_join, _cmd_dispatch_outcome
 if __package__:
     from .tickets_store import DEFAULT_RUN_STATE_TREE, NO_SINK_ERROR, RUN_STATE_TREES, _create_text_exclusively, _identity_update, _improvement_root, _load_ticket, _run_lock, _runs_root, _segment_error, _tickets_root, _write_identity, _write_text_atomically
@@ -416,7 +418,7 @@ def _dispatch(argv):
         from tickets import _sync_seams
     _sync_seams()
     if not argv:
-        return {'error': 'missing subcommand: new | lint | bound-check | instantiate | grade | gate | checker-stage | stamp-generation | draft-validate | seal | list | show | ready | dispatch | dispatch-open | dispatch-commit | dispatch-retire | dispatch-replace | dispatch-outcome | dispatch-join | dispatch-packet | dispatch-receive | check | set-status | join-noop-repair | result | worklog | run-state | improvement'}
+        return {'error': 'missing subcommand: new | lint | bound-check | instantiate | grade | gate | checker-stage | stamp-generation | draft-validate | seal | list | show | ready | dispatch | dispatch-open | dispatch-commit | dispatch-retire | dispatch-replace | dispatch-outcome | dispatch-join | dispatch-packet | dispatch-receive | dispatch-receipt | check | set-status | join-noop-repair | result | worklog | run-state | improvement'}
     command, rest = (argv[0], argv[1:])
     if command in HELP_COMMANDS:
         return _cmd_help()
@@ -461,6 +463,8 @@ def _dispatch(argv):
         return _cmd_dispatch_packet(rest)
     if command == 'dispatch-receive':
         return _cmd_dispatch_receive(rest)
+    if command == 'dispatch-receipt':
+        return _cmd_dispatch_receipt(rest)
     if command == 'check':
         return _cmd_check(rest)
     if command == 'set-status':
